Generate model for Food

+++ db/migrate/20220413065818_create_foods.rb
@@ -3,7 +3,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.text :description
-      t.integer :price
+      t.float :price
 
       t.timestamps
     end
